pyproject.toml
[project]
name = "actual-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "actualpy>=0.21.0",
]
